refactor(device-intent): move LWM LoadingContent to DeviceGenericStates
It sat inside the context initializer's components, so anything else showing a
DIE spinner had to reach into them. Desktop already keeps it in
DeviceGenericStates; mobile now matches.
