+++ apps/ledger-live-mobile/src/mvvm/components/DeviceIntentExecutor/components/DeviceGenericStates/LoadingContent.tsx
